# ZBP-polish-api-2_1_2-swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "2_1_2",
    "title" : "Polish API",
    "description" : "Interface specification for services provided by third parties based on access to payment accounts. Prepared by the Polish Bank Association and its affiliates"
  },
  "host" : "apiHost",
  "basePath" : "/",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
...
  "parameters" : {
...
    "acceptEncodingParam" : {
      "name" : "Accept-Encoding",
      "in" : "header",
      "required" : true,
      "type" : "string",
      "enum" : [ "gzip", "deflate" ],
      "description" : "Gzip, deflate"
    },
    "acceptLanguageParam" : {
      "name" : "Accept-Language",
      "in" : "header",
      "required" : true,
      "type" : "string",
      "description" : "Prefered language of response"
    },
    "acceptCharsetParam" : {
      "name" : "Accept-Charset",
      "in" : "header",
      "required" : true,
      "type" : "string",
      "enum" : [ "utf-8" ],
      "description" : "UTF-8"
    },
    "xjwsSignatureParam" : {
      "name" : "X-JWS-SIGNATURE",
      "in" : "header",
      "required" : true,
      "type" : "string",
      "description" : "Detached JWS signature of the body of the payload"
    },
    "xRequestIdParam" : {
      "name" : "X-REQUEST-ID",
      "in" : "header",
      "required" : true,
      "type" : "string",
      "description" : "Identyfikator żądania w formacie UUID (Wariant 1, Wersja 1), zgodnym ze standardem RFC 4122, nadawany przez TPP. Wartość musi być zgodna z parametrem requestId przekazywanym w ciele każdego żądania. / Request identifier using UUID format (Variant 1, Version 1), described in RFC 4122 standard, set by TPP. Value of this header must be the same as for the requestId param passed inside request payload."
    }
  },
  "paths" : {
    "/v2_1_2.1/auth/v2_1_2.1/authorize" : {
      "x-swagger-router-controller" : "as",
      "post" : {
        "summary" : "Requests OAuth2 authorization code",
        "description" : "Requests OAuth2 authorization code",
        "operationId" : "authorize",
        "parameters" : [ {
          "$ref" : "#/parameters/acceptEncodingParam"
        }, {
          "$ref" : "#/parameters/acceptLanguageParam"
        }, {
          "$ref" : "#/parameters/acceptCharsetParam"
        }, {
          "$ref" : "#/parameters/xjwsSignatureParam"
        }, {
          "$ref" : "#/parameters/xRequestIdParam"
        }, {
          "name" : "authorizeRequest",
          "in" : "body",
          "description" : "Data for OAuth2 Authorization Code Request",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/AuthorizeRequest"
          }
        } ],
        "tags" : [ "AS" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/AuthorizeResponse"
            },
            "headers" : {
              "Content-Encoding" : {
                "type" : "string",
                "description" : "Gzip, deflate"
              },
              "X-JWS-SIGNATURE" : {
                "type" : "string",
                "description" : "Detached JWS signature of the body of the response"
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "405" : {
            "description" : "Method Not Allowed",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "406" : {
            "description" : "Not Acceptable",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "415" : {
            "description" : "Unsupported Media Type",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "422" : {
            "description" : "Unprocessable entity",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "501" : {
            "description" : "Not Implemented",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "503" : {
            "description" : "Service Unavailable",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/v2_1_2.1/auth/v2_1_2.1/authorizeExt" : {
      "x-swagger-router-controller" : "as",
      "post" : {
        "summary" : "Requests OAuth2 authorization code based on One-time authorization code issued by External Authorization Tool",
        "description" : "Requests OAuth2 authorization code based One-time authorization code issued by External Authorization Tool. Authorization code will be delivered to TPP as callback request from ASPSP if PSU authentication is confirmed by EAT. Callback function must provide similar notification also in case of unsuccessful authentication or its abandonment.",
        "operationId" : "authorizeExt",
        "parameters" : [ {
          "$ref" : "#/parameters/acceptEncodingParam"
        }, {
          "$ref" : "#/parameters/acceptLanguageParam"
        }, {
          "$ref" : "#/parameters/acceptCharsetParam"
        }, {
          "$ref" : "#/parameters/xjwsSignatureParam"
        }, {
          "$ref" : "#/parameters/xRequestIdParam"
        }, {
          "name" : "eatCodeRequest",
          "in" : "body",
          "description" : "Data for OAuth2 Authorization Code Request extended for EAT based authentication and callback response",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/EatCodeRequest"
          }
        } ],
        "tags" : [ "AS" ],
        "responses" : {
          "204" : {
            "description" : "Successful TPP and EAT Code verification",
            "headers" : {
              "Content-Encoding" : {
                "type" : "string",
                "description" : "Gzip, deflate"
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "405" : {
            "description" : "Method Not Allowed",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "406" : {
            "description" : "Not Acceptable",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "415" : {
            "description" : "Unsupported Media Type",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "422" : {
            "description" : "Unprocessable entity",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "501" : {
            "description" : "Not Implemented",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "503" : {
            "description" : "Service Unavailable",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "xs2a_auth_aspsp" : {
      "type" : "oauth2",
      "flow" : "accessCode",
      "authorizationUrl" : "https://apiHost/v2_1_2.1/auth/v2_1_2.1/authorize",
      "tokenUrl" : "https://apiHost/v2_1_2.1/auth/v2_1_2.1/token",
      "scopes" : {
        "ais-accounts" : "Permission to retrieve the list of accounts through Account Information Service",
        "ais" : "Permission to execute methods of Account Information Service in context of particular account",
        "pis" : "Permission to initiate payments and check their statuses through Payment Initiation Service"
      }
    },
    "xs2a_auth_decoupled" : {
      "type" : "oauth2",
      "flow" : "accessCode",
      "authorizationUrl" : "https://apiHost/auth/authorizeExt",
      "tokenUrl" : "https://apiHost/v2_1_2.1/auth/v2_1_2.1/token",
      "scopes" : {
        "ais-accounts" : "Permission to retrieve the list of accounts through Account Information Service",
        "ais" : "Permission to execute methods of Account Information Service in context of particular account",
        "pis" : "Permission to initiate payments and check their statuses through Payment Initiation Service"
      }
    }
  },
  "definitions" : {
    "AuthorizeRequest" : {
      "description" : "Klasa zapytania o uzyskanie kodu autoryzacyjnego zgodnego z OAuth2",
      "type" : "object",
      "required" : [ "requestHeader", "response_type", "client_id", "redirect_uri", "scope", "scope_details", "state" ],
      "properties" : {
        "requestHeader" : {
          "$ref" : "#/definitions/RequestHeaderWithoutTokenAS"
        },
        "response_type" : {
          "type" : "string",
          "description" : "Typ odpowiedzi. Wartość stała: code"
        },
        "client_id" : {
          "type" : "string",
          "description" : "Identyfikator TPP"
        },
        "redirect_uri" : {
          "type" : "string",
          "description" : "Adres usługi TPP, na które zostanie wykonane przekierowanie po uwierzytelnieniu klienta ASPSP"
        },
        "scope" : {
          "type" : "string",
          "description" : "Typ zgody, o którą prosi TPP. Wybrana wartość z listy dostępnych identyfikatorów zgód, opisanych w specyfikacji standardu Polish API."
        },
        "scope_details" : {
          "$ref" : "#/definitions/ScopeDetailsInput"
        },
        "state" : {
          "type" : "string",
          "description" : "Losowa, unikalna w ramach TPP wartość – zabezpieczenie przed atakiem Cross-Site Request Forgery"
        }
      },
      "xml" : {
        "name" : "AuthorizeRequest"
      }
    },
    "AuthorizeResponse" : {
      "description" : "Klasa odpowiedzi na zapytanie TPP o autoryzację PSU do wykonania usługi interfejsu XS2A",
      "required" : [ "responseHeader", "aspspRedirectUri" ],
      "properties" : {
        "responseHeader" : {
          "$ref" : "#/definitions/ResponseHeader"
        },
        "aspspRedirectUri" : {
          "type" : "string",
          "description" : "Adres po stronie ASPSP, na który TPP powinien dokonać przekierowania przeglądarki PSU w celu jego uwierzytelnienia"
        }
      },
      "xml" : {
        "name" : "AuthorizeResponse"
      }
    },
    "EatCodeRequest" : {
      "description" : "Klasa zapytania o uzyskanie kodu autoryzacyjnego zgodnego z OAuth2 na podstawie przekazanego kodu jednorazowego, wygenerowanego w EAT",
      "type" : "object",
      "required" : [ "requestHeader", "requestId", "response_type", "eatCode", "client_id", "callbackURL", "apiKey", "scope", "scope_details" ],
      "properties" : {
        "requestHeader" : {
          "$ref" : "#/definitions/RequestHeaderWithoutTokenCallbackAS"
        },
        "response_type" : {
          "type" : "string",
          "description" : "Typ odpowiedzi. Wartość stała: code"
        },
        "eatCode" : {
          "type" : "string",
          "description" : "Jednorazowy kod autoryzacyjny wygenerowany przez EAT"
        },
        "eatType" : {
          "$ref" : "#/definitions/DictionaryItem"
        },
        "client_id" : {
          "type" : "string",
          "description" : "Identyfikator TPP"
        },
        "scope" : {
          "type" : "string",
          "description" : "Typ zgody, o którą prosi TPP. Wybrana wartość z listy dostępnych identyfikatorów zgód, opisanych w specyfikacji standardu Polish API."
        },
        "scope_details" : {
          "$ref" : "#/definitions/ScopeDetailsInput"
        },
        "state" : {
          "type" : "string",
          "description" : "Losowa, unikalna w ramach TPP wartość – zabezpieczenie przed atakiem Cross-Site Request Forgery"
        }
      },
      "xml" : {
        "name" : "EatCodeRequest"
      }
    },
...
    "Error" : {
      "description" : "Klasa zawierająca informacje o zwracanym błędzie / Error class",
      "properties" : {
        "responseHeader" : {
          "$ref" : "#/definitions/ResponseHeader"
        },
        "code" : {
          "type" : "string",
          "description" : "Kod błędu / Error code"
        },
        "message" : {
          "type" : "string",
          "description" : "Opis błędu / Error message"
        }
      }
    },
    "AccountNumber" : {
      "description" : "Numer konta",
      "type" : "string",
      "maxLength" : 34
    },
...
    "RequestHeaderWithoutToken" : {
      "description" : "Klasa zawierająca informacje o PSU / PSU Information Class",
      "type" : "object",
      "required" : [ "requestId" ],
      "properties" : {
        "requestId" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Identyfikator żądania w formacie UUID (Wariant 1, Wersja 1) zgodnym ze standardem RFC 4122 / Request ID using UUID format (Variant 1, Version 1) described in RFC 4122 standard"
        },
        "userAgent" : {
          "type" : "string",
          "description" : "Browser agent dla PSU / PSU browser agent"
        },
        "ipAddress" : {
          "type" : "string",
          "description" : "Adres IP końcowego urządzenia PSU. Wymagany dla isDirectPsu=true. / IP address of PSU's terminal device. Required when isDirectPsu=true."
        },
        "sendDate" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Oryginalna data wysłania, format: YYYY-MM-DDThh:mm:ss[.mmm] / Send date"
        },
        "tppId" : {
          "type" : "string",
          "description" : "Identyfikator TPP / TPP ID"
        }
      }
    },
    "RequestHeaderWithoutTokenAS" : {
      "description" : "Klasa zawierająca informacje o PSU na potrzeby usług autoryzacji / PSU Information Class dedicated for authorization services",
      "allOf" : [ {
        "$ref" : "#/definitions/RequestHeaderWithoutToken"
      }, {
        "type" : "object",
        "properties" : {
          "isCompanyContext" : {
            "type" : "boolean",
            "description" : "(true / false) Znacznik oznaczający czy żądanie jest wysyłane w kontekście PSU korporacyjnego"
          },
          "psuIdentifierType" : {
            "type" : "string",
            "description" : "Typ identyfikatora PSU, służy do wskazania przez TPP na podstawie jakiej informacji zostanie zidentyfikowany PSU, który będzie podlegał uwierzytelnieniu. Wartość słownikowa. / PSU identifier type, used by TPP to indicate type of information based on which the PSU is to be authenticated. Dictionary value."
          },
          "psuIdentifierValue" : {
            "type" : "string",
            "description" : "Wartość identyfikatora PSU. Wymagany warunkowo - w przypadku gdy została przekazana niepusta wartość typu identyfikatora PSU. / The value of the PSU's identifier. Required conditionally - in case non empty value of PSU identifier type was passed."
          },
          "psuContextIdentifierType" : {
            "type" : "string",
            "description" : "Typ identyfikatora kontekstu w jakim występuje PSU. Wartość słownikowa. Wymagany warunkowo - w przypadku wysyłania żądania dla takiego PSU, który może występować w więcej niż jednym kontekście w wybranym ASPSP. / Identifier of context that is used by PSU. Dictionary value. Required conditionally - in case context of the request is used and PSU may use more then one context for the ASPSP."
          },
          "psuContextIdentifierValue" : {
            "type" : "string",
            "description" : "Wartość identyfikatora kontekstu w jakim występuje PSU. Wymagany warunkowo - w przypadku wysyłania żądania dla takiego PSU, który może występować w więcej niż jednym kontekście w wybranym ASPSP. / The value of context that is used by PSU. Required conditionally - in case context of the request is used and PSU may use more then one context for the ASPSP."
          }
        }
      } ]
    },
    "RequestHeaderWithoutTokenCallbackAS" : {
      "description" : "Klasa zawierająca informacje o PSU and callback URL / PSU and callback URL Information Class",
      "allOf" : [ {
        "$ref" : "#/definitions/RequestHeaderWithoutTokenAS"
      }, {
        "type" : "object",
        "required" : [ "callbackURL, apiKey" ],
        "properties" : {
          "callbackURL" : {
            "type" : "string",
            "description" : "adres funkcji zwrotnej / callback URL"
          },
          "apiKey" : {
            "type" : "string",
            "description" : "API key dla wywołania funkcji zwrotnej / callback API key"
          }
        }
      } ]
    },
...
    "ResponseHeader" : {
      "description" : "Klasa zawierająca zwrotne metadane / Metadata Class",
      "properties" : {
        "requestId" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Identyfikator żądania otrzymanego od TPP / Identifier of request received from TPP"
        },
        "sendDate" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Data odpowiedzi, format: YYYY-MM-DDThh:mm:ss[.mmm] / Send date"
        },
        "isCallback" : {
          "type" : "boolean",
          "description" : "Znacznik określający czy odpowiedz zostanie przekazana w formie wywołania zwrotnego. true - gdy odpowiedz w formie wywołania zwrotnego. Inna dopuszczalna wartość to false."
        }
      }
    },
...
    "DictionaryItem" : {
      "description" : "Klasa zawierająca dane elementu/pozycji słownika / Dictionary Item Class",
      "properties" : {
        "code" : {
          "type" : "string",
          "description" : "Kod pozycji słownika / Code"
        },
        "description" : {
          "type" : "string",
          "description" : "Opis pozycji słownika / Description"
        }
      }
    },
    "ScopeDetailsInput" : {
      "description" : "Zakres, limity, parametry i czas ważności zgód, o które prosi TPP. / Scope, limits, parameters and expiration time of consents requested by TPP",
      "type" : "object",
      "required" : [ "consentId", "scopeGroupType", "scopeTimeLimit", "throttlingPolicy" ],
      "properties" : {
        "privilegeList" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ScopeDetailsInputPrivilegeList"
          },
          "description" : "The list of structures describing details of consent. Conditionally required - in case of process of obtaining new consent."
        },
        "scopeGroupType" : {
          "type" : "string",
          "description" : "Type of consent",
          "enum" : [ "ais-accounts", "ais", "pis" ]
        },
        "consentId" : {
          "type" : "string",
          "description" : "Id of consent"
        },
        "scopeTimeLimit" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Data ważności zgody / consent valid until, YYYY-MM-DDThh:mm:ss[.mmm]"
        },
        "throttlingPolicy" : {
          "type" : "string",
          "description" : "Throttling policy",
          "enum" : [ "psd2Regulatory" ]
        }
      }
    },
...
    "PrivilegeAisAspspInSimple" : {
      "description" : "Lista atrybutów uprawnienia usługi AIS będących przedmiotem zapytania o zgodę PSU / The list of attributes of the privilege of the AIS service that are the subject of the request for PSU's consent",
      "type" : "object",
      "properties" : {
        "scopeUsageLimit" : {
          "type" : "string",
          "enum" : [ "single", "multiple" ],
          "description" : "Rodzaj limitu zgody / Type of limit of usages"
        }
      }
    },
    "PrivilegeAisAspspIn" : {
      "description" : "Lista atrybutów uprawnienia usługi AIS będących przedmiotem zapytania o zgodę PSU / The list of attributes of the privilege of the AIS service that are the subject of the request for PSU's consent",
      "type" : "object",
      "required" : [ "maxAllowedHistoryLong" ],
      "properties" : {
        "scopeUsageLimit" : {
          "type" : "string",
          "enum" : [ "single", "multiple" ],
          "description" : "Rodzaj limitu zgody / Type of limit of usages"
        },
        "maxAllowedHistoryLong" : {
          "type" : "integer",
          "description" : "Ilość dni wstecz, liczona od momentu wysłania przez TPP żądania o pobranie historii transakcji rachunku płatnicznego, jaką może obejmować odpowiedź na to żądanie / How much account's transaction history is allowed to retrieve in days. This value should be used for calculation based on date of sending the request.",
          "minimum" : 1,
          "maximum" : 1460
        }
      }
    },
...
    "PrivilegePayment" : {
      "description" : "Lista atrybutów uprawnienia do żądania statusu płatności / The list of attributes of the payment status request privilege",
      "type" : "object",
      "properties" : {
        "scopeUsageLimit" : {
          "type" : "string",
          "enum" : [ "single", "multiple" ],
          "description" : "Rodzaj limitu zgody / Type of limit of usages"
        },
        "paymentId" : {
          "type" : "string",
          "description" : "Identyfikator płatności / Payment ID"
        },
        "tppTransactionId" : {
          "type" : "string",
          "description" : "Identyfikator transakcji nadany przez TPP. / Identifier of transaction established by TPP."
        }
      }
    },
    "PrivilegeBundle" : {
      "description" : "Lista atrybutów uprawnienia do żądania statusu paczki płatności / The list of attributes of the bundle of payments status request privilege",
      "type" : "object",
      "properties" : {
        "scopeUsageLimit" : {
          "type" : "string",
          "enum" : [ "single", "multiple" ],
          "description" : "Rodzaj limitu zgody / Type of limit of usages"
        },
        "bundleId" : {
          "type" : "string",
          "description" : "Identyfikator paczki płatności / Bundle of payments ID"
        },
        "tppBundleId" : {
          "type" : "string",
          "description" : "Identyfikator paczki przelewów nadany przez TPP. / Bundle of payments identifier set by TPP."
        }
      }
    },
...
    "PrivilegeDomesticTransfer" : {
      "description" : "Lista atrybutów uprawnienia do inicjacji przelewu zwykłego / The list of attributes of the domestic transfer request initiation privilege",
      "type" : "object",
      "properties" : {
        "scopeUsageLimit" : {
          "type" : "string",
          "enum" : [ "single", "multiple" ],
          "description" : "Rodzaj limitu zgody / Type of limit of usages"
        },
        "recipient" : {
          "$ref" : "#/definitions/RecipientPIS"
        },
        "sender" : {
          "$ref" : "#/definitions/SenderPISDomestic"
        },
        "transferData" : {
          "$ref" : "#/definitions/TransferData"
        },
        "deliveryMode" : {
          "description" : "Tryb pilności / Urgency mode",
          "type" : "string",
          "enum" : [ "ExpressD0", "StandardD1" ]
        },
        "system" : {
          "description" : "Droga jaką przelew ma być rozliczony / The way the transfer should be settled",
          "type" : "string",
          "enum" : [ "Elixir", "ExpressElixir", "Sorbnet", "BlueCash", "Internal" ]
        },
        "executionMode" : {
          "description" : "Tryb realizacji płatności. Nadrzędna informacja decydująca o tym w jakim trybie zostanie zrealizowana płatność. / Payment execution mode. The superior information deciding which mode is to be used to execute payment.",
          "type" : "string",
          "enum" : [ "Immediate", "FutureDated" ]
        }
      }
    },
    "PrivilegeForeignTransferEEA" : {
      "description" : "Lista atrybutów uprawnienia do inicjacji przelewu zagranicznego EEA / The list of attributes of the EEA foreign transfer request initiation privilege",
      "type" : "object",
      "properties" : {
        "scopeUsageLimit" : {
          "type" : "string",
          "enum" : [ "single", "multiple" ],
          "description" : "Rodzaj limitu zgody / Type of limit of usages"
        },
        "recipient" : {
          "$ref" : "#/definitions/RecipientPISForeign"
        },
        "sender" : {
          "$ref" : "#/definitions/SenderPISForeign"
        },
        "transferData" : {
          "$ref" : "#/definitions/TransferDataCurrencyRequired"
        },
        "deliveryMode" : {
          "description" : "Tryb pilności / Urgency mode",
          "type" : "string",
          "enum" : [ "ExpressD0", "StandardD1" ]
        },
        "system" : {
          "description" : "Droga jaką przelew ma być rozliczony / The way the transfer should be settled",
          "type" : "string",
          "enum" : [ "SEPA", "InstantSEPA", "Target" ]
        },
        "executionMode" : {
          "description" : "Tryb realizacji płatności. Nadrzędna informacja decydująca o tym w jakim trybie zostanie zrealizowana płatność. / Payment execution mode. The superior information deciding which mode is to be used to execute payment.",
          "type" : "string",
          "enum" : [ "Immediate", "FutureDated" ]
        }
      }
    },
    "PrivilegeForeignTransferNonEEA" : {
      "description" : "Lista atrybutów uprawnienia do inicjacji przelewu zagranicznego bez EEA / The list of attributes of the non-EEA foreign transfer request initiation privilege",
      "type" : "object",
      "properties" : {
        "scopeUsageLimit" : {
          "type" : "string",
          "enum" : [ "single", "multiple" ],
          "description" : "Rodzaj limitu zgody / Type of limit of usages"
        },
        "recipient" : {
          "$ref" : "#/definitions/RecipientPISForeign"
        },
        "recipientBank" : {
          "$ref" : "#/definitions/Bank"
        },
        "sender" : {
          "$ref" : "#/definitions/SenderPISForeign"
        },
        "transferData" : {
          "$ref" : "#/definitions/TransferDataCurrencyRequired"
        },
        "transferCharges" : {
          "type" : "string",
          "maxLength" : 3,
          "description" : "Klauzula kosztowa / The cost clause"
        },
        "deliveryMode" : {
          "description" : "Tryb pilności / Urgency mode",
          "type" : "string",
          "enum" : [ "ExpressD0", "UrgentD1", "StandardD2" ]
        },
        "system" : {
          "description" : "Droga jaką przelew ma być rozliczony / The way the transfer should be settled",
          "type" : "string",
          "enum" : [ "Swift" ]
        },
        "executionMode" : {
          "description" : "Tryb realizacji płatności. Nadrzędna informacja decydująca o tym w jakim trybie zostanie zrealizowana płatność. / Payment execution mode. The superior information deciding which mode is to be used to execute payment.",
          "type" : "string",
          "enum" : [ "Immediate", "FutureDated" ]
        }
      }
    },
...
    "ScopeDetailsInputPrivilegeList" : {
      "description" : "Pełna lista uprawnień będących przedmiotem zapytania o zgodę PSU / The list of all privileges that are the subject of the request for PSU's consent",
      "properties" : {
        "accountNumber" : {
          "$ref" : "#/definitions/AccountNumber"
        },
        "ais-accounts:getAccounts" : {
          "$ref" : "#/definitions/PrivilegeAisAspspInSimple"
        },
        "ais:getAccount" : {
          "$ref" : "#/definitions/PrivilegeAisAspspInSimple"
        },
        "ais:getHolds" : {
          "$ref" : "#/definitions/PrivilegeAisAspspIn"
        },
        "ais:getTransactionsDone" : {
          "$ref" : "#/definitions/PrivilegeAisAspspIn"
        },
        "ais:getTransactionsPending" : {
          "$ref" : "#/definitions/PrivilegeAisAspspIn"
        },
        "ais:getTransactionsRejected" : {
          "$ref" : "#/definitions/PrivilegeAisAspspIn"
        },
        "ais:getTransactionsCancelled" : {
          "$ref" : "#/definitions/PrivilegeAisAspspIn"
        },
        "ais:getTransactionsScheduled" : {
          "$ref" : "#/definitions/PrivilegeAisAspspIn"
        },
        "ais:getTransactionDetail" : {
          "$ref" : "#/definitions/PrivilegeAisAspspInSimple"
        },
        "pis:getPayment" : {
          "$ref" : "#/definitions/PrivilegePayment"
        },
        "pis:getBundle" : {
          "$ref" : "#/definitions/PrivilegeBundle"
        },
        "pis:domestic" : {
          "$ref" : "#/definitions/PrivilegeDomesticTransfer"
        },
        "pis:EEA" : {
          "$ref" : "#/definitions/PrivilegeForeignTransferEEA"
        },
        "pis:nonEEA" : {
          "$ref" : "#/definitions/PrivilegeForeignTransferNonEEA"
        },
        "pis:tax" : {
          "$ref" : "#/definitions/PrivilegeTaxTransfer"
        },
        "pis:cancelPayment" : {
          "$ref" : "#/definitions/PrivilegeCancelPayment"
        },
        "pis:bundle" : {
          "$ref" : "#/definitions/PrivilegeBundleTransfers"
        },
        "pis:recurring" : {
          "$ref" : "#/definitions/PrivilegeRecurringPayment"
        },
        "pis:getRecurringPayment" : {
          "$ref" : "#/definitions/PrivilegeRecurringPaymentStatus"
        },
        "pis:cancelRecurringPayment" : {
          "$ref" : "#/definitions/PrivilegeCancelRecurringPayment"
        }
      }
    },
...
  }
}